Track pedigree information for each animal
This commit closes #20. The animal model was modified to track the
progeny and parents of each animal. Those relationships are represented
on a pedigree page that includes links to the related animals. The
pedigree page also indicates who the animal's offspring were as well as
who their siblings were. On the generic information output, the
animal's type of birth (e.g. single, twin, triplet...) is recorded. A
toggle was added to indicate whether an animal was an orphan. This
information is used to update how the animal was raised (e.g. orphan,
single, twin, etc). Need to create a new issue to reflect stillborn
animals and animals that died before weaning into the "how raised"
count.

This commit also included an upgrade from bootstrap 3 to 4 closes #44