Update number of lambs weaned & number of lambings
This commit updates the progeny count to indicate the number of lambs
born and weaned and the number of lambings. The previous implementation
of the "raised as" number only incorporated the number of lambs born
minus the number of orphans. This version deducts the number of
stillborn lambs and those that are marked as "dead" at an age less than
40 days. This assumes a weaning age of 40 days and that animals would
not be sold or culled before a weaning age. Closes #45